requirements.txt
streamlit==1.28.0
openai==1.3.0
requests==2.31.0
pandas==2.1.0
numpy==1.25.0
faiss-cpu==1.7.4
sentence-transformers==2.7.0
huggingface_hub>=0.16.0,<1.0.0
transformers>=4.32.0
torch>=1.11.0
python-dotenv==1.0.0
tiktoken==0.5.1
langchain>=0.1.0,<0.2.0
langchain-core>=0.1.7,<0.2.0
langchain-community>=0.0.10,<0.1.0
langchain-text-splitters>=0.0.1,<0.1.0
